Improve AI logic
- don't visit stables if there's no reason for it
- prioritirize Gazebo
- allow to visit Lighthouse
- update building sequence for Sorceress castle
