+++ src/fheroes2/ai/normal/ai_normal_castle.cpp
@@ -62,6 +62,12 @@ namespace AI
                {DWELLING_MONSTER2, 3}, {DWELLING_MONSTER1, 4}, {BUILD_MAGEGUILD1, 2},  {BUILD_WEL2, 10},       {BUILD_TAVERN, 5},      {BUILD_THIEVESGUILD, 10},
                {BUILD_MAGEGUILD2, 3},  {BUILD_MAGEGUILD3, 4},  {BUILD_MAGEGUILD4, 5},  {BUILD_MAGEGUILD5, 5},  {BUILD_SHIPYARD, 4},    {BUILD_MARKETPLACE, 10}};
 
+        static const std::vector<BuildOrder> sorceressBuildOrder
+            = {{BUILD_CASTLE, 2},      {BUILD_STATUE, 1},      {DWELLING_MONSTER6, 1}, {DWELLING_MONSTER5, 1}, {DWELLING_MONSTER4, 1}, {BUILD_MAGEGUILD1, 1},
+               {DWELLING_MONSTER3, 1}, {DWELLING_UPGRADE4, 1}, {DWELLING_UPGRADE3, 2}, {DWELLING_UPGRADE2, 5}, {DWELLING_MONSTER2, 2}, {BUILD_TAVERN, 2},
+               {DWELLING_MONSTER1, 4}, {BUILD_WEL2, 10}, {BUILD_THIEVESGUILD, 10}, {BUILD_MAGEGUILD2, 3},  {BUILD_MAGEGUILD3, 4},  {BUILD_MAGEGUILD4, 5},
+               {BUILD_MAGEGUILD5, 5},  {BUILD_SHIPYARD, 4}, {BUILD_MARKETPLACE, 10}};
+
         // De-prioritizing dwelling 5 (you can reach 6 without it), 1 and upgrades of 3 and 4
         // Well, tavern and Archery upgrade are more important
         static const std::vector<BuildOrder> knightBuildOrder
@@ -92,6 +98,8 @@ namespace AI
             return necromancerBuildOrder;
         case Race::WZRD:
             return wizardBuildOrder;
+        case Race::SORC:
+            return sorceressBuildOrder;
         default:
             break;
         }
